+++ src/example-types/number.js
@@ -17,7 +17,54 @@ module.exports = {
       [context.field]: _.pickBy(_.isNumber, {
         gte: boundaryFilter(context.min),
         lte: boundaryFilter(context.max),
-      }),
-    },
+      })
+    }
   }),
+  result: async (context, search) => {
+    let field = _.get('field', context)
+    let result = await search({
+      aggs: {
+        statistical: {
+          stats: {
+            field,
+            missing: 0,
+          },
+        }
+      },
+    })
+
+    let statistical = _.get('aggregations.statistical', result)
+    let interval = Math.round(Math.abs(statistical.max - statistical.min) / 50) || 1
+    let histogram = []
+
+    if (interval) {
+      let valuesResult = await search({
+        aggs: {
+          values: {
+            histogram: {
+              field,
+              interval: interval,
+              min_doc_count: 0
+            }
+          }
+        },
+      })
+
+      let totalDocuments = _.get('hits.total', valuesResult)
+      histogram = _.map(entry => {
+        return {
+          value: Math.round(entry.key),
+          count: entry.doc_count
+        }
+      }, _.get('aggregations.values.buckets', valuesResult))
+    }
+
+    return {
+      results: {
+        interval,
+        statistical,
+        histogram
+      }
+    }
+  }
 }
